+++ addon/components/modal-dialog/component.js
@@ -8,39 +8,41 @@ export default class ModalDialogComponent extends Component {
   tagName = '';
   layout = layout;
 
+  // State
+
   isLoading = false;
   isShowing = true;
   isWarning = false;
 
+  // Arguments
+
   escapable = false;
 
-  onReady() {}
-  onLoaded() {}
-  onLoadError() {}
-  onClose() {}
+  // Actions
 
-  onLoad() {
-    return resolve();
-  }
+  onReady = null;
+  onLoaded = null;
+  onLoadError = null;
+  onClose = null;
+  onLoad = null;
 
-  constructor() {
-    super(...arguments);
+  init() {
+    super.init(...arguments);
     this.rootElement = document.querySelector(':root');
     this.documentElement = document.documentElement;
+    this._load();
   }
 
   @action
   close() {
-    return this._hide().then(() => this.onClose());
+    return this._hide().then(() => this._invokeAction('onClose'));
   }
 
   @action
   warn() {
     set(this, 'isWarning', true);
 
-    this._waitForAnimation().then(() => {
-      trySet(this, 'isWarning', false);
-    });
+    this._waitForAnimation().then(() => trySet(this, 'isWarning', false));
   }
 
   @action
@@ -62,7 +64,6 @@ export default class ModalDialogComponent extends Component {
     set(this, 'domElement', element);
     this.domElement.focus();
     this.rootElement.classList.add('has-modal');
-    this._load();
     this._watchForContentChanges();
     this._ready();
   }
@@ -78,7 +79,7 @@ export default class ModalDialogComponent extends Component {
   }
 
   _ready() {
-    this.onReady(this._api());
+    this._invokeAction('onReady', this._api());
   }
 
   _api() {
@@ -90,16 +91,10 @@ export default class ModalDialogComponent extends Component {
   _load() {
     set(this, 'isLoading', true);
 
-    this.onLoad()
-      .then(data => {
-        this.onLoaded(data);
-      })
-      .catch(error => {
-        this.onLoadError(error);
-      })
-      .finally(() => {
-        trySet(this, 'isLoading', false);
-      });
+    resolve(this._invokeAction('onLoad'))
+      .then(data => this._invokeAction('onLoaded', data))
+      .catch(error => this._invokeAction('onLoadError', error))
+      .finally(() => trySet(this, 'isLoading', false));
   }
 
   _hide() {
@@ -153,6 +148,12 @@ export default class ModalDialogComponent extends Component {
       this.close();
     } else {
       this.warn();
+    }
+  }
+
+  _invokeAction(name, ...args) {
+    if (typeof this[name] === 'function') {
+      return this[name](...args);
     }
   }
 }