+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 1.0.4
+
+- Move `@onLoad` action from did insert hook to component initialisation hook
+
+## 1.0.3
+
+- More resiliant if argument actions are not functions
+
 ## 1.0.2
 
 - Upgrade dependencies